Refactor CV into its own component - cv_section font is larger only in print layout
