+++ assets/css/jobs.scss
@@ -1,5 +1,4 @@
 .cv-section {
-  font-size: 1.1rem;
   margin: 1.75rem 0;
 
   .media-left {
@@ -18,12 +17,17 @@
       width: 76px;
     }
   }
-
   margin: 1.5rem 0;
 }
 
-div.onpage-qr-link {
-  position: absolute;
-  right: 0px;
-  bottom: -96px;
+.cv.is-print {
+  .cv-section {
+    font-size: 1.1rem;
+  }
+
+  div.onpage-qr-link {
+    position: absolute;
+    right: 0px;
+    bottom: -96px;
+  }
 }
