+++ lib/tmde_web/components/jobs/cv/cv_entry.html.heex
@@ -28,7 +28,7 @@
     <%= @description %>
     <ul>
       <%= for focus <- @entry.focuses do %>
-        <.cv_focus focus={focus} />
+        <.focus focus={focus} />
       <% end %>
     </ul>
   </div>
